homework_5 fix 2
